# apps/wallet-web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "relational-rust-server",
    "description": "Custodial Avalanche wallet service running inside Intel SGX enclave with Gramine.",
    "contact": {
      "name": "Relational Network"
    },
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/wallets/{wallet_id}/balance": {
      "get": {
        "tags": [
          "Wallets"
        ],
        "summary": "Get the balance of a wallet on the Avalanche C-Chain.",
        "description": "Returns native AVAX balance and any configured ERC-20 token balances.",
        "operationId": "get_wallet_balance",
        "parameters": [
          {
            "name": "wallet_id",
            "in": "path",
            "description": "Wallet ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "network",
            "in": "query",
            "description": "Network to query: \"fuji\" (testnet) or \"mainnet\"",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "default": "fuji"
            }
          },
          {
            "name": "tokens",
            "in": "query",
            "description": "Additional token contract addresses to query (comma-separated)",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Balance retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not wallet owner"
          },
          "404": {
            "description": "Wallet not found"
          },
          "503": {
            "description": "Blockchain network unavailable"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/wallets/{wallet_id}/balance/native": {
      "get": {
        "tags": [
          "Wallets"
        ],
        "summary": "Get only the native AVAX balance of a wallet (faster).",
        "operationId": "get_native_balance",
        "parameters": [
          {
            "name": "wallet_id",
            "in": "path",
            "description": "Wallet ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "network",
            "in": "query",
            "description": "Network to query: \"fuji\" (testnet) or \"mainnet\"",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "default": "fuji"
            }
          },
          {
            "name": "tokens",
            "in": "query",
            "description": "Additional token contract addresses to query (comma-separated)",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Native balance retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NativeBalanceResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not wallet owner"
          },
          "404": {
            "description": "Wallet not found"
          },
          "503": {
            "description": "Blockchain network unavailable"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BalanceResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/WalletBalanceResponse",
            "description": "Balance information"
          },
          {
            "type": "object",
            "required": [
              "wallet_id"
            ],
            "properties": {
              "wallet_id": {
                "type": "string",
                "description": "Wallet ID"
              }
            }
          }
        ],
        "description": "Balance response."
      },
...
      "NativeBalanceResponse": {
        "type": "object",
        "description": "Native token balance response (simpler version).",
        "required": [
          "wallet_id",
          "address",
          "network",
          "balance_wei",
          "balance"
        ],
        "properties": {
          "address": {
            "type": "string",
            "description": "Public address"
          },
          "balance": {
            "type": "string",
            "description": "AVAX balance formatted"
          },
          "balance_wei": {
            "type": "string",
            "description": "AVAX balance in wei"
          },
          "network": {
            "type": "string",
            "description": "Network name"
          },
          "wallet_id": {
            "type": "string",
            "description": "Wallet ID"
          }
        }
      },
...
      "TokenBalance": {
        "type": "object",
        "description": "Token balance information.",
        "required": [
          "symbol",
          "name",
          "balance_raw",
          "balance_formatted",
          "decimals"
        ],
        "properties": {
          "balance_formatted": {
            "type": "string",
            "description": "Balance formatted with decimals"
          },
          "balance_raw": {
            "type": "string",
            "description": "Balance in smallest unit (wei for native, token decimals for ERC-20)"
          },
          "contract_address": {
            "type": [
              "string",
              "null"
            ],
            "description": "Contract address (None for native token)"
          },
          "decimals": {
            "type": "integer",
            "format": "int32",
            "description": "Number of decimals",
            "minimum": 0
          },
          "name": {
            "type": "string",
            "description": "Token name"
          },
          "symbol": {
            "type": "string",
            "description": "Token symbol (e.g., \"AVAX\", \"EUROC\")"
          }
        }
      },
...
      "WalletBalanceResponse": {
        "type": "object",
        "description": "Wallet balance response including native and token balances.",
        "required": [
          "address",
          "network",
          "chain_id",
          "native_balance",
          "token_balances"
        ],
        "properties": {
          "address": {
            "type": "string",
            "description": "Wallet address"
          },
          "chain_id": {
            "type": "integer",
            "format": "int64",
            "description": "Chain ID",
            "minimum": 0
          },
          "native_balance": {
            "$ref": "#/components/schemas/TokenBalance",
            "description": "Native token balance (AVAX)"
          },
          "network": {
            "type": "string",
            "description": "Network name"
          },
          "token_balances": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TokenBalance"
            },
            "description": "ERC-20 token balances"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Clerk JWT token"
      }
    }
  },
...
}